patientId in check-in.component is now checked to make sure it is a number

+++ src/app/check-in/check-in.component.ts
@@ -19,13 +19,12 @@ export class CheckInComponent {
   constructor(private router: Router, private tsDataService: TsScreenerDataService, private tfiDataService: TfiDataService, private thsDataService: ThsDataService) {};
   
   public onClick() {
-    if(this.patientId.length === 4) {  
+    if(this.patientId.length === 4 && this.isPatientIdNumber()) {  
       sessionStorage.clear();
       this.tsDataService.clearHistory();
       this.tfiDataService.clearHistory();
       this.thsDataService.clearHistory();
       Utilities.setSessionStorage('patient-id', this.patientId);
-      Utilities.setSessionStorage('firstName', 'DEPRECATED: GET RID OF FIRST NAME!');
       console.log(this.patientId);
       this.router.navigateByUrl('/appointments');
       console.log('log in with ' + this.patientId);
@@ -36,7 +35,19 @@ export class CheckInComponent {
     }
   }
   
-  
+  isPatientIdNumber() : boolean {
+    let result : boolean = true;
+
+    let i : number = 0;
+    for(i = 0; i < this.patientId.length; i++) {
+      if(isNaN(parseInt(this.patientId[i], 10))) {
+        result = false;
+        break;
+      }
+    }
+
+    return result;
+  }
   
   /**
    * This function is called when the "enter" key on the keyboard is pressed.
@@ -44,7 +55,7 @@ export class CheckInComponent {
    * @param event the event caught by the action.
   */
   public keyDownFunction(event) {
-    if ( event.keyCode === 13) {
+    if (event.keyCode === 13) {
       this.onClick();
     } else {
       this.authenticationFlag = true;
